personal-finance-tracker/requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
opencv-python
pillow
google-generativeai
